+++ packages/go_router/example/lib/others/nav_builder.dart
@@ -6,7 +6,27 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:provider/provider.dart';
 
-import '../shared/data.dart';
+/// The login information.
+class LoginInfo extends ChangeNotifier {
+  /// The username of login.
+  String get userName => _userName;
+  String _userName = '';
+
+  /// Whether a user has logged in.
+  bool get loggedIn => _userName.isNotEmpty;
+
+  /// Logs in a user.
+  void login(String userName) {
+    _userName = userName;
+    notifyListeners();
+  }
+
+  /// Logs out the current user.
+  void logout() {
+    _userName = '';
+    notifyListeners();
+  }
+}
 
 void main() => runApp(App());
 
@@ -35,28 +55,7 @@ class App extends StatelessWidget {
         name: 'home',
         path: '/',
         builder: (BuildContext context, GoRouterState state) =>
-            HomeScreenNoLogout(families: Families.data),
-        routes: <GoRoute>[
-          GoRoute(
-            name: 'family',
-            path: 'family/:fid',
-            builder: (BuildContext context, GoRouterState state) {
-              final Family family = Families.family(state.params['fid']!);
-              return FamilyScreen(family: family);
-            },
-            routes: <GoRoute>[
-              GoRoute(
-                name: 'person',
-                path: 'person/:pid',
-                builder: (BuildContext context, GoRouterState state) {
-                  final Family family = Families.family(state.params['fid']!);
-                  final Person person = family.person(state.params['pid']!);
-                  return PersonScreen(family: family, person: person);
-                },
-              ),
-            ],
-          ),
-        ],
+            const HomeScreenNoLogout(),
       ),
       GoRoute(
         name: 'login',
@@ -66,39 +65,23 @@ class App extends StatelessWidget {
       ),
     ],
 
-    // redirect to the login page if the user is not logged in
-    redirect: (GoRouterState state) {
-      // if the user is not logged in, they need to login
-      final bool loggedIn = _loginInfo.loggedIn;
-      final String loginloc = state.namedLocation('login');
-      final bool loggingIn = state.subloc == loginloc;
-
-      // bundle the location the user is coming from into a query parameter
-      final String homeloc = state.namedLocation('home');
-      final String fromloc = state.subloc == homeloc ? '' : state.subloc;
-      if (!loggedIn) {
-        return loggingIn
-            ? null
-            : state.namedLocation(
-                'login',
-                queryParams: <String, String>{
-                  if (fromloc.isNotEmpty) 'from': fromloc
-                },
-              );
-      }
-
-      // if the user is logged in, send them where they were going before (or
-      // home if they weren't going anywhere)
-      if (loggingIn) {
-        return state.queryParams['from'] ?? homeloc;
-      }
-
-      // no need to redirect at all
-      return null;
-    },
-
     // changes on the listenable will cause the router to refresh it's route
     refreshListenable: _loginInfo,
+
+    // redirect to the login page if the user is not logged in
+    redirect: (GoRouterState state) {
+      final bool loggedIn = _loginInfo.loggedIn;
+      const String loginLocation = '/login';
+      final bool loggingIn = state.subloc == loginLocation;
+
+      if (!loggedIn) {
+        return loggingIn ? null : loginLocation;
+      }
+      if (loggingIn) {
+        return state.namedLocation('home');
+      }
+      return null;
+    },
 
     // add a wrapper around the navigator to:
     // - put loginInfo into the widget tree, and to
@@ -108,7 +91,6 @@ class App extends StatelessWidget {
             ChangeNotifierProvider<LoginInfo>.value(
       value: _loginInfo,
       builder: (BuildContext context, Widget? _) {
-        debugPrint('navigatorBuilder: ${state.subloc}');
         return _loginInfo.loggedIn ? AuthOverlay(child: child) : child;
       },
     ),
@@ -145,67 +127,14 @@ class AuthOverlay extends StatelessWidget {
 /// The home screen without a logout button.
 class HomeScreenNoLogout extends StatelessWidget {
   /// Creates a [HomeScreenNoLogout].
-  const HomeScreenNoLogout({required this.families, Key? key})
-      : super(key: key);
-
-  /// The list of families.
-  final List<Family> families;
+  const HomeScreenNoLogout({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) => Scaffold(
         appBar: AppBar(title: const Text(App.title)),
-        body: ListView(
-          children: <Widget>[
-            for (final Family f in families)
-              ListTile(
-                title: Text(f.name),
-                onTap: () => context
-                    .goNamed('family', params: <String, String>{'fid': f.id}),
-              )
-          ],
+        body: const Center(
+          child: Text('home screen'),
         ),
-      );
-}
-
-/// The screen that shows a list of persons in a family.
-class FamilyScreen extends StatelessWidget {
-  /// Creates a [FamilyScreen].
-  const FamilyScreen({required this.family, Key? key}) : super(key: key);
-
-  /// The family to display.
-  final Family family;
-
-  @override
-  Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(title: Text(family.name)),
-        body: ListView(
-          children: <Widget>[
-            for (final Person p in family.people)
-              ListTile(
-                title: Text(p.name),
-                onTap: () => context.go('/family/${family.id}/person/${p.id}'),
-              ),
-          ],
-        ),
-      );
-}
-
-/// The person screen.
-class PersonScreen extends StatelessWidget {
-  /// Creates a [PersonScreen].
-  const PersonScreen({required this.family, required this.person, Key? key})
-      : super(key: key);
-
-  /// The family this person belong to.
-  final Family family;
-
-  /// The person to be displayed.
-  final Person person;
-
-  @override
-  Widget build(BuildContext context) => Scaffold(
-        appBar: AppBar(title: Text(person.name)),
-        body: Text('${person.name} ${family.name} is ${person.age} years old'),
       );
 }