+++ packages/go_router/example/test/path_and_query_params_test.dart
@@ -24,8 +24,8 @@ void main() {
     await tester.pumpAndSettle();
     // 'Chris' should be higher than 'Tom'.
     expect(
-        tester.getCenter(find.text('Chris')).dy <
-            tester.getCenter(find.text('Tom')).dy,
+        tester.getCenter(find.text('Jane')).dy <
+            tester.getCenter(find.text('John')).dy,
         isTrue);
 
     testRouteInformation = <String, dynamic>{
@@ -40,8 +40,8 @@ void main() {
     await tester.pumpAndSettle();
     // 'Chris' should be lower than 'Tom'.
     expect(
-        tester.getCenter(find.text('Chris')).dy >
-            tester.getCenter(find.text('Tom')).dy,
+        tester.getCenter(find.text('Jane')).dy >
+            tester.getCenter(find.text('John')).dy,
         isTrue);
   });
 }